multi app fixes

+++ opinion-infrastructure/src/main/java/com/inqwise/opinion/infrastructure/common/Cookie.java
@@ -117,7 +117,7 @@ public class Cookie implements ICookie {
     }
 
     @Override
-    public Integer getMaxAge() {
+    public long getMaxAge() {
         return maxAge;
     }
