+++ opinion-infrastructure/src/main/java/com/inqwise/opinion/infrastructure/common/ICookie.java
@@ -33,7 +33,7 @@ public interface ICookie {
 	
 	String getDomain();
 
-	Integer getMaxAge();
+	long getMaxAge();
 
 	String getName();
